# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for checking wallet security and calculating nutrition score",
        "title": "Wallet Nutrition Score API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/check/{address}": {
            "get": {
                "description": "Check wallet security and get nutrition score",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "Check wallet security",
                "parameters": [
                    {
                        "type": "string",
                        "format": "eth_addr",
                        "description": "Wallet address to check (must be valid Ethereum address starting with 0x)",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.CheckWalletResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Check if the service is running",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.CheckResult": {
            "type": "object",
            "properties": {
                "check_name": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "raw_data": {},
                "risk_found": {
                    "type": "boolean"
                },
                "risk_level": {
                    "$ref": "#/definitions/entity.RiskLevel"
                },
                "score_penalty": {
                    "type": "number"
                }
            }
        },
        "entity.RiskLevel": {
            "type": "string",
            "enum": [
                "LOW",
                "MEDIUM",
                "HIGH",
                "CRITICAL"
            ],
            "x-enum-varnames": [
                "RiskLevelLow",
                "RiskLevelMedium",
                "RiskLevelHigh",
                "RiskLevelCritical"
            ]
        },
        "main.CheckWalletResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "checks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.CheckResult"
                    }
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "recommendations": {
                    "type": "string"
                },
                "score": {
                    "type": "number"
                }
            }
        }
    }
}